+++ lib/src/code_excerpt_updater.dart
@@ -12,6 +12,8 @@ import 'package:yaml/yaml.dart';
 import 'util.dart';
 import 'nullable.dart';
 
+const defaultPlaster = '···';
+
 const _eol = '\n';
 Function _listEq = const ListEquality().equals;
 typedef String CodeTransformer(String code);
@@ -183,6 +185,8 @@ class Updater {
             infoPath,
             info.region,
             [
+              plasterCodeTransformer(args['plaster'],
+                  _determineCodeLang(openingCodeBlockLine, info.path)),
               removeCodeTransformer(args['remove']),
               retainCodeTransformer(args['retain']),
               replaceCodeTransformer(args['replace']),
@@ -265,7 +269,7 @@ class Updater {
   }
 
   RegExp supportedArgs = new RegExp(
-      r'^(class|diff-with|from|indent-by|path-base|region|replace|remove|retain|title|to)$');
+      r'^(class|diff-with|from|indent-by|path-base|plaster|region|replace|remove|retain|title|to)$');
 
   void __extractAndNormalizeNamedArgs(InstrInfo info, String argsAsString) {
     if (argsAsString == null) return;
@@ -324,7 +328,7 @@ class Updater {
     return result;
   }
 
-  /*@nullable*/
+  @nullable
   Iterable<String> _getDiff(String relativeSrcPath1, Map<String, String> args) {
     final relativeSrcPath2 = args['diff-with'];
     final pathPrefix = p.join(srcDirPath, _pathBase);
@@ -405,7 +409,7 @@ class Updater {
     return '${match[1]} $path';
   }
 
-  /*@nullable*/
+  @nullable
   Iterable<String> _getExcerpt(
       String relativePath, String region, CodeTransformer t) {
     String excerpt = _getExcerptAsString(relativePath, region);
@@ -514,7 +518,52 @@ class Updater {
   final _matchDollarNumRE = new RegExp(r'(\$+)(&|\d*)');
   final _endRE = new RegExp(r'^g;?\s*$');
 
-  /*@nullable*/
+  /// Replace raw plaster markers in excerpt with [plasterTemplate].
+  /// Note that plaster line indentation is not affected.
+  ///
+  /// If [plasterTemplate] is 'none' then plasters are removed.
+  /// If [plasterTemplate] is null then a default [lang] specific plaster
+  /// template is used.
+  @nullable
+  CodeTransformer plasterCodeTransformer(String plasterTemplate, String lang) {
+    if (plasterTemplate == 'none') return removeCodeTransformer(defaultPlaster);
+    if (!excerptsYaml) return null;
+
+    final template =
+        plasterTemplate?.replaceAll(r'$defaultPlaster', defaultPlaster) ??
+            _plasterTemplateFor(lang);
+    return template == null
+        ? null
+        : _replaceCodeTransformer(defaultPlaster, template);
+  }
+
+  @nullable
+  String _plasterTemplateFor(String lang) {
+    if (lang == null) return null;
+
+    switch (lang) {
+      case 'css':
+        return '/* $defaultPlaster */';
+
+      case 'html':
+        return '<!-- $defaultPlaster -->';
+
+      case 'dart':
+      case 'js':
+      case 'scss':
+      case 'ts':
+        return '// $defaultPlaster';
+
+      case 'yaml':
+        return '# $defaultPlaster';
+
+      case 'diff':
+      default:
+        return null;
+    }
+  }
+
+  @nullable
   CodeTransformer replaceCodeTransformer(String replaceExp) {
     dynamic _reportErr([String extraInfo = '']) =>
         _reportError('invalid replace attribute ("$replaceExp"); ' +
@@ -556,7 +605,7 @@ class Updater {
     return transformers.fold(null, compose);
   }
 
-  /*@nullable*/
+  @nullable
   CodeTransformer _replaceCodeTransformer(String re, String _replacement) {
     final replacement = encodeSlashChar(_replacement);
     _log.finest(' >> replacement expr: $replacement');
@@ -586,7 +635,7 @@ class Updater {
             }));
   }
 
-  /*@nullable*/
+  @nullable
   CodeTransformer removeCodeTransformer(String arg) {
     final Predicate<String> matcher = _retainArgToMatcher('remove', arg);
     return matcher == null
@@ -594,7 +643,7 @@ class Updater {
         : _lineMatcherToCodeTransformer(_not(matcher));
   }
 
-  /*@nullable*/
+  @nullable
   CodeTransformer retainCodeTransformer(String arg) {
     final Predicate<String> matcher = _retainArgToMatcher('retain', arg);
     return matcher == null ? null : _lineMatcherToCodeTransformer(matcher);
@@ -631,6 +680,17 @@ class Updater {
 
   void _report(String prefix, String msg) =>
       _stderr.writeln('$prefix: $_filePath:$lineNum $msg');
+
+  final RegExp _codeBlockLangSpec = new RegExp(r'(?:```|prettify\s+)(\w+)');
+
+  String _determineCodeLang(String openingCodeBlockLine, String path) {
+    final match = _codeBlockLangSpec.firstMatch(openingCodeBlockLine);
+    if (match != null) return match[1];
+
+    var ext = p.extension(path);
+    if (ext.startsWith('.')) ext = ext.substring(1);
+    return ext;
+  }
 }
 
 class InstrInfo {